Every single time I come around

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -12,7 +12,7 @@
 
 char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
-	unsigned int len1 = 0, len2 = 0, l;
+	size_t len1 = 0, len2 = 0, l;
 	char *ptr;
 
 	if (s1 == NULL)
